+++ apps/backoffice-v2/src/lib/blocks/hooks/useProcessingDetailsBlock/useProcessingDetailsBlock.tsx
@@ -1,5 +1,6 @@
 import { useMemo } from 'react';
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
+import { userCreatedIconCell } from '@/lib/blocks/utils/constants';
 
 export const useProcessingDetailsBlock = ({ processingDetails, workflow }) => {
   return useMemo(() => {
@@ -14,12 +15,29 @@ export const useProcessingDetailsBlock = ({ processingDetails, workflow }) => {
         value: createBlocksTyped()
           .addBlock()
           .addCell({
-            type: 'heading',
-            value: 'Processing details',
-          })
-          .addCell({
-            type: 'subheading',
-            value: 'User-Provided Data',
+            type: 'container',
+            value: createBlocksTyped()
+              .addBlock()
+              .addCell(userCreatedIconCell)
+              .addCell({
+                type: 'container',
+                value: createBlocksTyped()
+                  .addBlock()
+                  .addCell({
+                    type: 'heading',
+                    value: 'Processing details',
+                    props: { className: 'mt-0' },
+                  })
+                  .addCell({
+                    type: 'subheading',
+                    value: 'User-Provided Data',
+                  })
+                  .buildFlat(),
+              })
+              .buildFlat(),
+            props: {
+              className: 'flex space-x-1 items-center mt-4',
+            },
           })
           .addCell({
             type: 'details',