+++ services/workflows-service/src/assessments/assessments.module.ts
@@ -1,7 +1,7 @@
+import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { AssessmentsService } from './assessments.service';
 import { CustomerModule } from '@/customer/customer.module';
-import { HttpModule } from '@nestjs/axios';
 import { AssessmentsControllerExternal } from './assessments.controller.external';
 import { UnifiedApiClient } from '@/common/utils/unified-api-client/unified-api-client';