+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -2023,8 +2023,7 @@ export class WorkflowService {
       return;
     }
 
-    this.sentry.captureException(new Error('Workflow definition context validation failed'));
-    this.logger.error('Workflow definition context validation failed', {
+    this.logger.warn('Workflow definition context validation failed', {
       errors: validate.errors,
       errorData: validate.errors?.map(error => ({
         path: error.instancePath,
@@ -2731,7 +2730,13 @@ export class WorkflowService {
       transaction,
     );
     const document = runtimeData?.context?.documents?.find(
-      (document: DefaultContextSchema['documents'][number]) => document.id === documentId,
+      (document: DefaultContextSchema['documents'][number]) => {
+        if (document?._document?.id) {
+          return document._document?.id === documentId;
+        }
+
+        return document.id === documentId;
+      },
     );
 
     return addPropertiesSchemaToDocument(document, workflowDef.documentsSchema);