+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/blocks
 
+## 0.2.46
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.105
+
+## 0.2.45
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.103
+
+## 0.2.44
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.101
+
 ## 0.2.43
 
 ### Patch Changes